+++ src/main/resources/application.properties
@@ -1,10 +1,7 @@
 spring.application.name=demoSpringBoot
-spring.datasource.url=jdbc:mysql://localhost:3306/mydatabase
-spring.datasource.username=root
-spring.datasource.password=password
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.jpa.hibernate.ddl-auto=create-drop
-spring.jpa.show-sql=true
 spring.devtools.restart.enabled=false
 security.jwt.secret-key=3cfa76ef14937c1c0ea519f8fc057a80fcd04a7420f8e8bcd0a7567c272e007b
 security.jwt.expiration-time=3600000
+
+spring.data.mongodb.uri=mongodb://admin:password@localhost:27017/yourdatabase?authSource=admin
+spring.data.mongodb.database=db
